src/requirements.txt
streamlit
Pillow
numpy
python-dotenv
opencv-python-headless==4.10.0.84
pandas

libgl1
libglib2.0-0
libsm6
libxext6
libxrender1



